+++ examples/02_search-for-accounts.rs
@@ -3,7 +3,7 @@
 //! Run the following Cargo command to run this example:
 //!
 //! ```bash
-//! cargo run --example 02_search-for-account
+//! cargo run --example 02_search-for-accounts
 //! ```
 
 use tetr_ch::prelude::*;
@@ -13,10 +13,9 @@ async fn main() {
     let client = Client::new();
 
     // Set the social connection to search for.
-    // The API document says searching for the other social links will be added in the near future.
-    let social_connection = SocialConnection::Discord("724976600873041940".to_string());
+    let social_connection = SocialConnection::RedditUsername("Rinrin0413".to_string());
 
-    // Search for the account.
+    // Search for the accounts.
     let response = match client.search_user(social_connection).await {
         Ok(res) => res,
         Err(err) => panic!("Response error: {}\n", err),
@@ -26,9 +25,11 @@ async fn main() {
         panic!("Error: {}\n", err.msg.expect("no error message"));
     }
 
-    let data = response.data.unwrap().user.unwrap();
-    println!("Name: {}", data.username);
-    println!("ID: {}", data.id);
+    let data = response.data.unwrap();
+    for user in data.users {
+        println!("Name: {}", user.username);
+        println!("ID: {}", user.id);
+    }
 
     // For more information about the data structure, see:
     // https://docs.rs/tetr_ch/latest/tetr_ch/model/searched_user/struct.UserData.html